renamed realEstate to a generic "item"
