+++ listopia-react/src/config/categories.js
@@ -1,0 +1,1 @@
+export default ['Apartment', 'House', 'Room', 'Land', 'Other'];
