+++ listopia-react/src/reducers/index.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
-import realEstates from './real-estates';
+import items from './items';
 import user from './user';
 import category from './category';
 
-export default combineReducers({ user, realEstates, category });
+export default combineReducers({ user, items, category });
